requirements.txt
typer>=0.9.0
questionary>=2.0.0
rich>=13.0.0
prompt_toolkit>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
langchain_core>=0.2.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0
anthropic>=0.18.0
chromadb>=0.4.0
httpx>=0.25.0
aiosqlite>=0.19.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
